pyproject.toml
[project]
name = "multi-doc-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.25",
    "llama-parse>=0.6.89",
    "python-dotenv>=1.2.1",
]
